+++ tests/testthat/test-createArea.R
@@ -22,8 +22,8 @@ sapply(studies, function(study) {
   
   
   test_that("Backup study/input", {
-    expect_length(backupStudy(what = "study"), 1)
-    expect_length(backupStudy(what = "input"), 1)
+    expect_length(backupStudy(what = "study", extension = ".zip"), 1)
+    expect_length(backupStudy(what = "input", extension = ".tar.gz"), 1)
   })